+++ R/convert_stat_chisq.R
@@ -8,6 +8,10 @@
 #' @param nrow,ncol The number of rows/columns in the contingency table (ignored
 #'   for Phi when `adjust=FALSE` and `CI=NULL`).
 #' @param ci Confidence Interval (CI) level
+#' @param alternative a character string specifying the alternative hypothesis;
+#'   Controls the type of CI returned: `"greater"` (default) or `"less"`
+#'   (one-sided CI), or `"two.sided"` (default, two-sided CI). Partial matching
+#'   is allowed (e.g., `"g"`, `"l"`, `"two"`...).
 #' @param adjust Should the effect size be bias-corrected? Defaults to `FALSE`.
 #' @param ... Arguments passed to or from other methods.
 #'
@@ -36,7 +40,7 @@
 #' #
 #' #         Pearson's Chi-squared test
 #' #
-#' # data:  ctab
+#' # data:  contingency_table
 #' # X-squared = 41.234, df = 4, p-value = 2.405e-08
 #'
 #' chisq_to_phi(41.234,
@@ -58,7 +62,8 @@
 #' Journal of the Korean Statistical Society, 42(3), 323-328.
 #'
 #' @export
-chisq_to_phi <- function(chisq, n, nrow, ncol, ci = 0.9, adjust = FALSE, ...) {
+chisq_to_phi <- function(chisq, n, nrow, ncol, ci = 0.95, alternative = "greater", adjust = FALSE, ...) {
+  alternative <- match.arg(alternative, c("two.sided", "less", "greater"))
   if (adjust || is.numeric(ci)) {
     is_goodness <- ncol == 1 || nrow == 1
 
@@ -82,12 +87,13 @@ chisq_to_phi <- function(chisq, n, nrow, ncol, ci = 0.9, adjust = FALSE, ...) {
   if (is.numeric(ci)) {
     stopifnot(length(ci) == 1, ci < 1, ci > 0)
     res$CI <- ci
+    ci.level <- if (alternative == "two.sided") ci else 2 * ci - 1
 
     chisq_ <- phi_to_chisq(res[[1]], n)
 
     chisqs <- t(mapply(
       .get_ncp_chi,
-      chisq_, df, ci
+      chisq_, df, ci.level
     ))
 
     res$CI_low <-
@@ -96,12 +102,20 @@ chisq_to_phi <- function(chisq, n, nrow, ncol, ci = 0.9, adjust = FALSE, ...) {
       chisq_to_phi(chisqs[, 2], n, nrow, ncol, ci = NULL, adjust = FALSE)[[1]]
 
     ci_method <- list(method = "ncp", distribution = "chisq")
+    if (alternative == "less") {
+      res$CI_low <- 0
+    } else if (alternative == "greater") {
+      res$CI_high <- 1
+    }
+  } else {
+    alternative <- NULL
   }
 
   class(res) <- c("effectsize_table", "see_effectsize_table", class(res))
   attr(res, "ci") <- ci
   attr(res, "ci_method") <- ci_method
   attr(res, "adjust") <- adjust
+  attr(res, "alternative") <- alternative
   return(res)
 }
 
@@ -113,7 +127,8 @@ chisq_to_cohens_w <- chisq_to_phi
 
 #' @rdname chisq_to_phi
 #' @export
-chisq_to_cramers_v <- function(chisq, n, nrow, ncol, ci = 0.9, adjust = FALSE, ...) {
+chisq_to_cramers_v <- function(chisq, n, nrow, ncol, ci = 0.95, alternative = "greater", adjust = FALSE, ...) {
+  alternative <- match.arg(alternative, c("two.sided", "less", "greater"))
   is_goodness <- ncol == 1 || nrow == 1
 
 
@@ -150,12 +165,13 @@ chisq_to_cramers_v <- function(chisq, n, nrow, ncol, ci = 0.9, adjust = FALSE, .
   if (is.numeric(ci)) {
     stopifnot(length(ci) == 1, ci < 1, ci > 0)
     res$CI <- ci
+    ci.level <- if (alternative == "two.sided") ci else 2 * ci - 1
 
     chisq_ <- phi_to_chisq(phi, n)
 
     chisqs <- t(mapply(
       .get_ncp_chi,
-      chisq_, df, ci
+      chisq_, df, ci.level
     ))
 
     res$CI_low <-
@@ -164,12 +180,20 @@ chisq_to_cramers_v <- function(chisq, n, nrow, ncol, ci = 0.9, adjust = FALSE, .
       chisq_to_cramers_v(chisqs[, 2], n, nrow, ncol, ci = NULL, adjust = FALSE)[[1]]
 
     ci_method <- list(method = "ncp", distribution = "chisq")
+    if (alternative == "less") {
+      res$CI_low <- 0
+    } else if (alternative == "greater") {
+      res$CI_high <- 1
+    }
+  } else {
+    alternative <- NULL
   }
 
   class(res) <- c("effectsize_table", "see_effectsize_table", class(res))
   attr(res, "ci") <- ci
   attr(res, "ci_method") <- ci_method
   attr(res, "adjust") <- adjust
+  attr(res, "alternative") <- alternative
   return(res)
 }