+++ R/convert_stat_to_anova.R
@@ -128,25 +128,25 @@
 #' Psychological Methods, 9, 164-182.
 #'
 #' @export
-F_to_eta2 <- function(f, df, df_error, ci = 0.9, ...) {
-  .F_to_pve(f, df, df_error, ci = ci, es = "eta2")
+F_to_eta2 <- function(f, df, df_error, ci = 0.95, alternative = "greater", ...) {
+  .F_to_pve(f, df, df_error, ci = ci, alternative = alternative, es = "eta2")
 }
 
 #' @rdname F_to_eta2
 #' @export
-t_to_eta2 <- function(t, df_error, ci = 0.9, ...) {
+t_to_eta2 <- function(t, df_error, ci = 0.95, alternative = "greater", ...) {
   F_to_eta2(t^2, 1, df_error, ci = ci)
 }
 
 #' @rdname F_to_eta2
 #' @export
-F_to_epsilon2 <- function(f, df, df_error, ci = 0.9, ...) {
-  .F_to_pve(f, df, df_error, ci = ci, es = "epsilon2")
+F_to_epsilon2 <- function(f, df, df_error, ci = 0.95, alternative = "greater", ...) {
+  .F_to_pve(f, df, df_error, ci = ci, alternative = alternative, es = "epsilon2")
 }
 
 #' @rdname F_to_eta2
 #' @export
-t_to_epsilon2 <- function(t, df_error, ci = 0.9, ...) {
+t_to_epsilon2 <- function(t, df_error, ci = 0.95, alternative = "greater", ...) {
   F_to_epsilon2(t^2, 1, df_error, ci = ci)
 }
 
@@ -160,13 +160,13 @@ t_to_eta2_adj <- t_to_epsilon2
 
 #' @rdname F_to_eta2
 #' @export
-F_to_omega2 <- function(f, df, df_error, ci = 0.9, ...) {
-  .F_to_pve(f, df, df_error, ci = ci, es = "omega2")
+F_to_omega2 <- function(f, df, df_error, ci = 0.95, alternative = "greater", ...) {
+  .F_to_pve(f, df, df_error, ci = ci, alternative = alternative, es = "omega2")
 }
 
 #' @rdname F_to_eta2
 #' @export
-t_to_omega2 <- function(t, df_error, ci = 0.9, ...) {
+t_to_omega2 <- function(t, df_error, ci = 0.95, alternative = "greater", ...) {
   F_to_omega2(t^2, 1, df_error, ci = ci)
 }
 
@@ -174,7 +174,7 @@ t_to_omega2 <- function(t, df_error, ci = 0.9, ...) {
 #' @rdname F_to_eta2
 #' @param squared Return Cohen's *f* or Cohen's *f*-squared?
 #' @export
-F_to_f <- function(f, df, df_error, ci = 0.9, squared = FALSE, ...) {
+F_to_f <- function(f, df, df_error, ci = 0.95, alternative = "greater", squared = FALSE, ...) {
   res_eta <- F_to_eta2(f, df, df_error, ci = ci)
 
   res <- data.frame(
@@ -206,25 +206,27 @@ F_to_f <- function(f, df, df_error, ci = 0.9, squared = FALSE, ...) {
 
 #' @rdname F_to_eta2
 #' @export
-t_to_f <- function(t, df_error, ci = 0.9, squared = FALSE, ...) {
-  F_to_f(t^2, 1, df_error, ci = ci, squared = squared)
+t_to_f <- function(t, df_error, ci = 0.95, alternative = "greater", squared = FALSE, ...) {
+  F_to_f(t^2, 1, df_error, ci = ci, alternative = alternative, squared = squared)
 }
 
 #' @rdname F_to_eta2
 #' @export
-F_to_f2 <- function(f, df, df_error, ci = 0.9, squared = TRUE, ...) {
-  F_to_f(f, df = df, df_error = df_error, ci = ci, squared = squared)
+F_to_f2 <- function(f, df, df_error, ci = 0.95, alternative = "greater", squared = TRUE, ...) {
+  F_to_f(f, df = df, df_error = df_error, ci = ci, alternative = alternative, squared = squared)
 }
 
 #' @rdname F_to_eta2
 #' @export
-t_to_f2 <- function(t, df_error, ci = 0.9, squared = TRUE, ...) {
-  F_to_f(t^2, 1, df_error, ci = ci, squared = squared)
+t_to_f2 <- function(t, df_error, ci = 0.95, alternative = "greater", squared = TRUE, ...) {
+  F_to_f(t^2, 1, df_error, ci = ci, alternative = alternative, squared = squared)
 }
 
 
 #' @keywords internal
-.F_to_pve <- function(f, df, df_error, ci = 0.9, es = "eta2") {
+.F_to_pve <- function(f, df, df_error, ci = 0.95, alternative = "greater", es = "eta2") {
+  alternative <- match.arg(alternative, c("two.sided", "less", "greater"))
+
   res <- switch(tolower(es),
     eta2 = data.frame(Eta2_partial = (f * df) / (f * df + df_error)),
     epsilon2 = data.frame(Epsilon2_partial = ((f - 1) * df) / (f * df + df_error)),
@@ -236,19 +238,29 @@ t_to_f2 <- function(t, df_error, ci = 0.9, squared = TRUE, ...) {
   if (is.numeric(ci)) {
     stopifnot(length(ci) == 1, ci < 1, ci > 0)
     res$CI <- ci
+    ci.level <- if (alternative == "two.sided") ci else 2 * ci - 1
+
     # based on MBESS::ci.R2
     f <- pmax(0, (res[[1]] / df) / ((1 - res[[1]]) / df_error))
-    fs <- t(mapply(.get_ncp_F, f, df, df_error, ci))
+    fs <- t(mapply(.get_ncp_F, f, df, df_error, ci.level))
 
     # This really is a generic F_to_R2
     res$CI_low <- F_to_eta2(fs[, 1], df, df_error, ci = NULL)[[1]]
     res$CI_high <- F_to_eta2(fs[, 2], df, df_error, ci = NULL)[[1]]
 
     ci_method <- list(method = "ncp", distribution = "F")
+    if (alternative == "less") {
+      res$CI_low <- 0
+    } else if (alternative == "greater") {
+      res$CI_high <- 1
+    }
+  } else {
+    alternative <- NULL
   }
 
   class(res) <- c("effectsize_table", "see_effectsize_table", class(res))
   attr(res, "ci") <- ci
   attr(res, "ci_method") <- ci_method
+  attr(res, "alternative") <- alternative
   return(res)
 }