+++ R/print.effectsize_table.R
@@ -10,6 +10,17 @@
 print.effectsize_table <- function(x, digits = 2, ...) {
   x_orig <- x
 
+  if (!is.null(alt <- attr(x, "alternative")) && alt != "two.sided") {
+    ci_footer <- sprintf(
+      "\n- One-sided CIs: %s bound fixed at (%s).",
+      if (alt == "less") "lower" else "upper",
+      as.character(if (alt == "less") x$CI_low[1] else x$CI_high[1])
+    )
+
+    attr(x, "table_footer") <-
+      c(attr(x, "table_footer"), list(c(ci_footer, "cyan")))
+  }
+
   x <- format(x, digits = digits)
 
   cat(insight::export_table(x, digits = digits, ...))