+++ R/xtab.R
@@ -6,6 +6,11 @@
 #'
 #' @inheritParams stats::chisq.test
 #' @param ci Confidence Interval (CI) level
+#' @param alternative a character string specifying the alternative hypothesis;
+#'   Controls the type of CI returned: `"two.sided"` (two-sided CI; default for
+#'   Cramer's *V*, phi (\eqn{\phi}), and Cohen's *w*), `"greater"` (default for
+#'   OR, RR, Cohen's *h* and Cohen's *g*) or `"less"` (one-sided CI). Partial
+#'   matching is allowed (e.g., `"g"`, `"l"`, `"two"`...).
 #' @param adjust Should the effect size be bias-corrected? Defaults to `FALSE`.
 #' @param ... Arguments passed to [stats::chisq.test()], such as `p`. Ignored
 #'   for `cohens_g()`.
@@ -112,12 +117,14 @@
 #'
 #' @importFrom stats chisq.test
 #' @export
-phi <- function(x, y = NULL, ci = 0.9, adjust = FALSE, ...) {
+phi <- function(x, y = NULL, ci = 0.95, alternative = "greater", adjust = FALSE, ...) {
+  alternative <- match.arg(alternative, c("two.sided", "less", "greater"))
+
   if (inherits(x, "BFBayesFactor")) {
     if (!inherits(x@numerator[[1]], "BFcontingencyTable")) {
       stop("'x' is not a Chi-squared test!", call. = FALSE)
     }
-    return(effectsize(x, type = "phi", adjust = adjust, ci = ci))
+    return(effectsize(x, type = "phi", adjust = adjust, ci = ci, alternative = alternative))
   }
 
 
@@ -131,7 +138,7 @@ phi <- function(x, y = NULL, ci = 0.9, adjust = FALSE, ...) {
     x$data.name <- NULL
   }
 
-  effectsize(x, type = "phi", adjust = adjust, ci = ci)
+  effectsize(x, type = "phi", adjust = adjust, ci = ci, alternative = alternative)
 }
 
 #' @rdname phi
@@ -141,12 +148,14 @@ cohens_w <- phi
 #' @rdname phi
 #' @importFrom stats chisq.test
 #' @export
-cramers_v <- function(x, y = NULL, ci = 0.9, adjust = FALSE, ...) {
+cramers_v <- function(x, y = NULL, ci = 0.95, alternative = "greater", adjust = FALSE, ...) {
+  alternative <- match.arg(alternative, c("two.sided", "less", "greater"))
+
   if (inherits(x, "BFBayesFactor")) {
     if (!inherits(x@numerator[[1]], "BFcontingencyTable")) {
       stop("'x' is not a Chi-squared test!", call. = FALSE)
     }
-    return(effectsize(x, type = "cramers_v", adjust = adjust, ci = ci))
+    return(effectsize(x, type = "cramers_v", adjust = adjust, ci = ci, alternative = alternative))
   }
 
 
@@ -160,7 +169,7 @@ cramers_v <- function(x, y = NULL, ci = 0.9, adjust = FALSE, ...) {
     x$data.name <- NULL
   }
 
-  effectsize(x, type = "cramers_v", adjust = adjust, ci = ci)
+  effectsize(x, type = "cramers_v", adjust = adjust, ci = ci, alternative = alternative)
 }
 
 
@@ -168,13 +177,15 @@ cramers_v <- function(x, y = NULL, ci = 0.9, adjust = FALSE, ...) {
 #' @inheritParams oddsratio_to_d
 #' @export
 #' @importFrom stats chisq.test qnorm
-oddsratio <- function(x, y = NULL, ci = 0.95, log = FALSE, ...) {
+oddsratio <- function(x, y = NULL, ci = 0.95, alternative = "two.sided", log = FALSE, ...) {
+  alternative <- match.arg(alternative, c("two.sided", "less", "greater"))
+
   if (inherits(x, "htest")) {
     if (grepl("Pearson's Chi-squared", x$method) ||
       grepl("Chi-squared test for given probabilities", x$method)) {
-      return(effectsize(x, type = "or", log = log, ci = ci))
+      return(effectsize(x, type = "or", log = log, ci = ci, alternative = alternative))
     } else if (grepl("Fisher's Exact", x$method)) {
-      return(effectsize(x, ...))
+      return(effectsize(x, alternative = alternative, ...))
     } else {
       stop("'x' is not a Chi-squared / Fisher's Exact test!", call. = FALSE)
     }
@@ -205,8 +216,9 @@ oddsratio <- function(x, y = NULL, ci = 0.95, log = FALSE, ...) {
   if (is.numeric(ci)) {
     stopifnot(length(ci) == 1, ci < 1, ci > 0)
     res$CI <- ci
+    ci.level <- if (alternative == "two.sided") ci else 2 * ci - 1
 
-    alpha <- 1 - ci
+    alpha <- 1 - ci.level
 
     SE_logodds <- sqrt(sum(1 / Obs))
     Z_logodds <- stats::qnorm(alpha / 2, lower.tail = FALSE)
@@ -216,6 +228,13 @@ oddsratio <- function(x, y = NULL, ci = 0.95, log = FALSE, ...) {
     res$CI_high <- confs[2]
 
     ci_method <- list(method = "normal")
+    if (alternative == "less") {
+      res$CI_low <- 0
+    } else if (alternative == "greater") {
+      res$CI_high <- Inf
+    }
+  } else {
+    alternative <- NULL
   }
 
   if (log) {
@@ -229,6 +248,7 @@ oddsratio <- function(x, y = NULL, ci = 0.95, log = FALSE, ...) {
   attr(res, "ci_method") <- ci_method
   attr(res, "log") <- log
   attr(res, "approximate") <- FALSE
+  attr(res, "alternative") <- alternative
   return(res)
 }
 
@@ -236,13 +256,15 @@ oddsratio <- function(x, y = NULL, ci = 0.95, log = FALSE, ...) {
 #' @inheritParams oddsratio_to_d
 #' @export
 #' @importFrom stats chisq.test qnorm
-riskratio <- function(x, y = NULL, ci = 0.95, log = FALSE, ...) {
+riskratio <- function(x, y = NULL, ci = 0.95, alternative = "two.sided", log = FALSE, ...) {
+  alternative <- match.arg(alternative, c("two.sided", "less", "greater"))
+
   if (inherits(x, "htest")) {
     if (!(grepl("Pearson's Chi-squared", x$method) ||
       grepl("Chi-squared test for given probabilities", x$method))) {
       stop("'x' is not a Chi-squared test!", call. = FALSE)
     }
-    return(effectsize(x, type = "rr", log = log, ci = ci))
+    return(effectsize(x, type = "rr", log = log, ci = ci, alternative = alternative))
   } else if (inherits(x, "BFBayesFactor")) {
     if (!inherits(x@numerator[[1]], "BFcontingencyTable")) {
       stop("'x' is not a Chi-squared test!", call. = FALSE)
@@ -273,8 +295,9 @@ riskratio <- function(x, y = NULL, ci = 0.95, log = FALSE, ...) {
   if (is.numeric(ci)) {
     stopifnot(length(ci) == 1, ci < 1, ci > 0)
     res$CI <- ci
+    ci.level <- if (alternative == "two.sided") ci else 2 * ci - 1
 
-    alpha <- 1 - ci
+    alpha <- 1 - ci.level
 
     SE_logRR <- sqrt(p1 / ((1 - p1) * n1)) + sqrt(p2 / ((1 - p2) * n2))
     Z_logRR <- stats::qnorm(alpha / 2, lower.tail = FALSE)
@@ -284,6 +307,13 @@ riskratio <- function(x, y = NULL, ci = 0.95, log = FALSE, ...) {
     res$CI_high <- confs[2]
 
     ci_method <- list(method = "normal")
+    if (alternative == "less") {
+      res$CI_low <- 0
+    } else if (alternative == "greater") {
+      res$CI_high <- Inf
+    }
+  } else {
+    alternative <- NULL
   }
 
   if (log) {
@@ -297,19 +327,22 @@ riskratio <- function(x, y = NULL, ci = 0.95, log = FALSE, ...) {
   attr(res, "ci_method") <- ci_method
   attr(res, "log") <- log
   attr(res, "approximate") <- FALSE
+  attr(res, "alternative") <- alternative
   return(res)
 }
 
 #' @rdname phi
 #' @export
 #' @importFrom stats qnorm
-cohens_h <- function(x, y = NULL, ci = 0.95, ...) {
+cohens_h <- function(x, y = NULL, ci = 0.95, alternative = "two.sided", ...) {
+  alternative <- match.arg(alternative, c("two.sided", "less", "greater"))
+
   if (inherits(x, "htest")) {
     if (!(grepl("Pearson's Chi-squared", x$method) ||
       grepl("Chi-squared test for given probabilities", x$method))) {
       stop("'x' is not a Chi-squared test!", call. = FALSE)
     }
-    return(effectsize(x, type = "cohens_h", ci = ci))
+    return(effectsize(x, type = "cohens_h", ci = ci, alternative = alternative))
   } else if (inherits(x, "BFBayesFactor")) {
     if (!inherits(x@numerator[[1]], "BFcontingencyTable")) {
       stop("'x' is not a Chi-squared test!", call. = FALSE)
@@ -339,22 +372,31 @@ cohens_h <- function(x, y = NULL, ci = 0.95, ...) {
   ci_method <- NULL
   if (is.numeric(ci)) {
     stopifnot(length(ci) == 1, ci < 1, ci > 0)
+    out$CI <- ci
+    ci.level <- if (alternative == "two.sided") ci else 2 * ci - 1
 
-    alpha <- (1 - ci) / 2
+    alpha <- 1 - ci.level
 
     se_arcsin <- sqrt(0.25 * (1 / n1 + 1 / n2))
-    Zc <- stats::qnorm(alpha, lower.tail = FALSE)
-    out$CI <- ci
+    Zc <- stats::qnorm(alpha / 2, lower.tail = FALSE)
     out$CI_low <- H - Zc * (2 * se_arcsin)
     out$CI_high <- H + Zc * (2 * se_arcsin)
 
     ci_method <- list(method = "normal")
+    if (alternative == "less") {
+      out$CI_low <- 0
+    } else if (alternative == "greater") {
+      out$CI_high <- Inf
+    }
+  } else {
+    alternative <- NULL
   }
 
   class(out) <- c("effectsize_table", "see_effectsize_table", class(out))
   attr(out, "ci") <- ci
   attr(out, "ci_method") <- ci_method
   attr(out, "approximate") <- FALSE
+  attr(out, "alternative") <- alternative
   return(out)
 }
 
@@ -362,12 +404,14 @@ cohens_h <- function(x, y = NULL, ci = 0.95, ...) {
 #' @rdname phi
 #' @export
 #' @importFrom stats complete.cases prop.test
-cohens_g <- function(x, y = NULL, ci = 0.95, ...) {
+cohens_g <- function(x, y = NULL, ci = 0.95, alternative = "two.sided", ...) {
+  alternative <- match.arg(alternative, c("two.sided", "less", "greater"))
+
   if (inherits(x, "htest")) {
     if (!grepl("McNemar", x$method)) {
       stop("'x' is not a McNemar test!", call. = FALSE)
     }
-    return(effectsize(x, ci = ci))
+    return(effectsize(x, ci = ci, alternative = alternative))
   }
 
 
@@ -402,11 +446,15 @@ cohens_g <- function(x, y = NULL, ci = 0.95, ...) {
 
   ci_method <- NULL
   if (is.numeric(ci)) {
+    stopifnot(length(ci) == 1, ci < 1, ci > 0)
+    out$CI <- ci
+
     n <- sum(b) + sum(c)
     k <- P * n
 
     res <- stats::prop.test(k, n,
       p = 0.5,
+      alternative = alternative,
       conf.level = ci,
       correct = FALSE
     )
@@ -422,5 +470,6 @@ cohens_g <- function(x, y = NULL, ci = 0.95, ...) {
   attr(out, "ci") <- ci
   attr(out, "ci_method") <- ci_method
   attr(out, "approximate") <- FALSE
+  attr(out, "alternative") <- alternative
   return(out)
 }